acme: change color scheme
A color-impoverished black and gray color scheme:

- Main window: light gray text on black background.
- Tag bar: light gray text on medium gray background.
- Text highlighting: shades of gray.

acme2k [1] was used as a reference on changing the the Acme color
scheme.  However, rather than putting the colors in config.h macros
like acme2k does, I changed them directly in acme.c.  This results in
slightly less divergence from upstream Acme.

Some of the colors being used came from this screenshot [2] but many
were changed.  The shade of gray used for the text in the main window
is the same as Konsole's "Linux Colors" color scheme.

[1] https://github.com/karahobny/acme2k
[2] https://raw.githubusercontent.com/karahobny/acme2k/master/img/a1.png