acme: don't have both Ctrl-<key> and Command-<key>
Make the Ctrl-[cvxzys] and Command-[cvxzZs] key bindings mutually
exclusive.  On macOS, this avoids creating unnecessary bindings that
could conflict with other uses (e.g., Ctrl-c for interrupt).  On
generic Unix, this doesn't change much, except remove a tiny amount
of dead code.

As a side effect, this change allows this fork of plan9port to be
used as an override source tree for the plan9port nixpkg, e.g.:

  environment.systemPackages = [
    (pkgs.plan9port.overrideAttrs (oldAttrs: {
      name = "plan9port-custom";
      src = /home/me/src/plan9port;
    }))
  ];

This did not work before, since the plan9port nixpkg makes the (rather
opinionated) decision to patch-in the Ctrl-[cvxz] key bindings.  See:

https://github.com/NixOS/nixpkgs/blob/695b3515251873e0a7e2021add4bba643c56cde3/pkgs/tools/system/plan9port/default.nix#L39

Since this tree already has a similar change (since commit 072fc2fa54),
the switch statements wind up with duplicate case values (a compiler
error).  With this commit, the substitution command in the plan9port
nixpkg no longer matches anything: this leads to some harmless warnings
when building the plan9port derivation, but it works.